Fix: don't force coupled nodes unready when ancestors unlistened if node still listening.
Normally a listening node's descendants wouldn't be listening.  But there's a race condition:  if a child
is listening, and an ancestor starts listening, then both remain listening until the ancestor receives its
first snapshot.  If the ancestor is unlistened before this happens, we need to recognize that the child is
still ready, otherwise we end up with a listening node that never becomes ready (unless there's a change
in its data that causes a new snapshot to arrive).