When planting a null or undefined value, delete the property instead like Firebase would.
While _applyLocalWrite already prunes the tree appropriately if given a top-level null or
undefined value, if such a value is nested inside a more complex object it gets cached in
the tree as-is.  This usually doesn't matter, but can be confusing to code that's using
a null to delete a subtree in Firebase, but then finds the local value to be null instead
of undefined.

We apply the same logic to local subtrees since it seems reasonable to keep the semantics
consistent with remote ones.