Add an `{immediate: true}` option to `observe`. (#22)
All observers are immediate in the Vue sense of the term, but we normally
delay the first callback until we've had a chance to return the unwatch
function to the user, since some observers will want to cancel themselves
on the first run.  However, this breaks some tightly timed observers that
need to be really, actually immediate, and don't care about unwatching.

This change repurposes the `immediate` option (which wasn't valid in
Firetruss before) to bypass the first callback delay mechanism.