Fix edge case when `onSlow` is applied to reads.
I think it's possible for a read to trigger `onSlow`, then
run `clearReady` before the operation ever actually became
ready.  This would result in consecutive calls to the `onSlow`
callback with `ready` set to `false`, breaking the contract
and throwing off any counts the callback was keeping.  Instead,
fake-set `ready` to `true` before resetting the `onSlow`
handlers since we're starting a new round of slowness detection
anyway.

This didn't affect Reviewable as we only listen to `onSlow` for
write operations, which never `clearReady`.